src/requirements.txt
cloudshell-shell-core~=5.0
cloudshell-cp-core~=2.0
